The cubes now spin when their button is selected and show an image on the back.

+++ SSAGOCodebreakers/Assets/Scripts/Grid Handling/GridDraw.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GridDraw : MonoBehaviour
 {
-    public int x_Start, y_Start;
+    public float x_Start, y_Start;
 
     public int width, height;
